requirements.txt
streamlit
pypdf
python-docx
redlines
beautifulsoup4
markdown